deleting the StartUpPhaseCommandParser.java to have a generic CommandParser class
