+++ src/main/java/game/commands/StartUpCommandValidator.java
@@ -1,4 +1,9 @@
 package game.commands;
 
-public class StartUpCommandValidator {
+public class StartUpCommandValidator implements CommandValidator{
+    @Override
+    public boolean validate(Command command) throws IllegalArgumentException{
+        return true;
+    }
+
 }
